feat(geo): add WhoIs/Contact entities with ipinfo.io integration and shared test fixtures (#37)
  Add comprehensive geographic entity support for IP geolocation (WhoIs) and contact
  management with content-addressable storage, nested location enrichment, and optimized
  test infrastructure.

  ## Domain Layer
  - Add WhoIs entity with SHA-256 content-addressable hashing from IP+year+month+fingerprint
  - Add Contact entity with ContextKey/RelatedEntityId for loose coupling to external entities
  - Add value objects: Personal, Professional, ContactMethods, EmailAddress, PhoneNumber
  - Implement temporal versioning for WhoIs to track IP location changes over time

  ## Application Layer
  - Add CQRS query handlers: GetLocationsByIds, GetWhoIsByIds, GetContactsByIds, GetContactsByExtKeys
  - Add CQRS command handlers: CreateLocations, CreateWhoIs, CreateContacts, DeleteContacts
  - Add FindWhoIs complex handler with cache check → external API → persist flow
  - Implement memory cache → repository fallback pattern with SOME_FOUND/NOT_FOUND propagation
  - Enrich WhoIs/Contact responses with nested Location data when present
  - Add GetMany/SetMany in-memory cache handlers for batch operations

  ## Infrastructure Layer
  - Add repository handlers using Batch.Pg for efficient batched queries
  - Add EF Core configurations for WhoIs and Contact entities
  - Implement IIpInfoClient abstraction wrapping ipinfo.io SDK for testability
  - Add Populate handler for hydrating WhoIs from external API responses

  ## gRPC API
  - Add endpoints: FindWhoIs, GetContacts, GetContactsByExtKeys, CreateContacts, DeleteContacts
  - Keep GetWhoIsByIds as internal-only (external consumers use FindWhoIs)

  ## Test Infrastructure
  - Add SharedPostgresFixture with ICollectionFixture<T> pattern
  - Reduce container startup overhead from 14 containers to 3 (~75% reduction)
  - Add comprehensive integration tests for all handlers with location enrichment scenarios

  ## Documentation
  - Update GEO_APP.md, GEO_INFRA.md, GEO_DOMAIN.md, GEO_TESTS.md

+++ backends/Services/Geo/Geo.Domain/ValueObjects/Coordinates.cs
@@ -29,7 +29,7 @@ public record Coordinates
     /// <example>
     /// 48.42841.
     /// </example>
-    public required decimal Latitude { get; init; }
+    public required double Latitude { get; init; }
 
     /// <summary>
     /// Gets longitude in decimal degrees.
@@ -37,7 +37,7 @@ public record Coordinates
     /// <example>
     /// -123.36564.
     /// </example>
-    public required decimal Longitude { get; init; }
+    public required double Longitude { get; init; }
 
     #region Functionality
 
@@ -60,8 +60,8 @@ public record Coordinates
     /// Thrown if latitude or longitude are out of range.
     /// </exception>
     public static Coordinates Create(
-        decimal latitude,
-        decimal longitude)
+        double latitude,
+        double longitude)
     {
         if (latitude is < -90 or > 90)
         {
